Explore: Facing, Sound, Know Position Update
Player Units now face the correct direction after attacking
Sound now alerts enemies when attacking and when another enemy spots a
player unit.
Known Position changed back to an arrayList. Further development is
needed.

It's a god-awful small affair
To the girl with the mousy hair
But her mummy is yelling "No"
And her daddy has told her to go
But her friend is nowhere to be seen