Explore: New Attack Method, Explore Map Extends INTO THE ABYSS
Thou shalt see further on, an Abyss was begat of the ancient beast, and
threatens to swallow the whole of Oolacile. Knight Artorias came to stop
this, but such a hero has nary a murmur of Dark. Without doubt he will
be swallowed by the Abyss, overcome by its utter blackness... Indeed,
the Abyss may be unstoppable..